+++ History.md
@@ -1,4 +1,17 @@
 
+2.1.6 / 2016-02-01
+==================
+
+  * Fix failed wait with queued up functions
+  * fix fullscreen switching (#434)
+
+2.1.5 / 2016-02-01
+==================
+
+  * add .insert(selector[, text]).
+  * improve .type(selector[, text]) robustness.
+  * bump electron and fix API updates.
+
 2.1.4 / 2016-01-28
 ==================